+++ example/example_task_download_file.dart
@@ -65,14 +65,12 @@ Task<String> _download(Uri uri, String filename, CancellationToken token,
 
     token.throwIfCanceled();
     _message('Starting download');
-    _message('Fetching feed: $uri');
     final request = Request('GET', uri);
     final task = Task.run(() => Client().send(request));
     StreamedResponse response;
     try {
       response = await task.withCancellation(token);
     } on TaskCanceledException {
-      // Ignore the cancelled connection establishment.
       unawaited(() async {
         try {
           await (await task).stream.listen((_) {}).cancel();